Homework - somehow my commit got messed up last time. Fixing test resources and a few minor fixes.

+++ Program1/ThreadedClient/src/edu/nku/common/LogicalClock.java
@@ -15,7 +15,7 @@ public class LogicalClock {
 		currentTime += sleepTime;
 	}
 	
-	public synchronized void decideIncrement(int otherTime){
+	public synchronized void decideIncrement(double otherTime){
 		currentTime = Math.max(currentTime + 1, otherTime + 1);
 	}
